+++ db/migrate/20140326154307_add_public_to_tools.rb
@@ -1,0 +1,5 @@
+class AddPublicToTools < ActiveRecord::Migration
+  def change
+    add_column :tools, :public, :boolean, null: false, default: true
+  end
+end