+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Dataset editors can now approve user access to datasets based on DAUA criteria
 - Removed variable lists to clean up interface
   - Variables are now downloaded via the dataset CSV download from the files area
+- Datasets can now specify release dates to allow users to see when those datasets will have files available
 - **Gem Changes**
   - Updated to rails 4.0.4
   - Updated to carrierwave 0.10.0